netdog: Add `generate-hostname` subcommand
This adds a new subcommand `generate-hostname` that is meant to be used
as a settings generator.  It will attempt to read the current IP from
file and resolve it via DNS reverse lookup.  If the lookup is successful
it is used, otherwise the IP is used with dots replaced by dashes to
avoid any confusion from other software that may attempt to read it, e.g
"1.2.3.4" will end up being hostname "ip-1-2-3-4".